requirements.txt
beautifulsoup4==4.12.2
selenium==4.14.0
Django==4.2.6
djangorestframework==3.14.0